+++ app/routes/v1/form/routes.py
@@ -6,3 +6,4 @@ from .files import *
 from .misc import *
 from .responses import *
 from .api import *
+from .analytics import *